requirements/base.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.32
idna==3.6
nodeenv==1.8.0
packaging==23.2
Pillow==9.5.0
platformdirs==4.0.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.5.5
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.24.7
